pyproject.toml
[project]
name = "diagrams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "daft>=0.1.2",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "pyro-ppl>=1.9.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
]
